requirement.txt
langchain
langchain.community
langchain-text-splitters
sentence-transformers
python-dotenv
langchain-chroma
langchain_huggingface
